pyproject.toml
[tool.poetry]
name = "alg-vandringen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ipykernel = "^6.29.4"
opencv-python = "^4.9.0.80"
selenium = "^4.19.0"
ultralytics = "^8.2.2"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
